Make the dynamic lazy attr adder be a Mixin.
